+++ CesiumRasterOverlays/include/CesiumRasterOverlays/AzureMapsRasterOverlay.h
@@ -218,20 +218,6 @@ public:
       CesiumUtility::IntrusivePointer<const RasterOverlay> pOwner)
       const override;
 
-  /**
-   * @brief Refresh a previously-created tile provider using a new key.
-   *
-   * Calling this method on a tile provider that was not created by this @ref
-   * AzureMapsRasterOverlay will lead to undefined behavior.
-   *
-   * @param pProvider The previously-created tile provider.
-   * @param newKey The new key to use.
-   */
-  CesiumAsync::Future<void> refreshTileProviderWithNewKey(
-      const CesiumUtility::IntrusivePointer<RasterOverlayTileProvider>&
-          pProvider,
-      const std::string& newKey);
-
 private:
   AzureMapsSessionParameters _sessionParameters;
 };